Split Filters+Strings files;
